+++ specs/src/main/java/io/nuun/kernel/api/predicates/ClassDescendantOf.java
@@ -14,38 +14,29 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with Nuun IO Kernel Specs.  If not, see <http://www.gnu.org/licenses/>.
  */
-package io.nuun.kernel.api.config;
+package io.nuun.kernel.api.predicates;
 
+import java.util.function.Predicate;
 
-public class KernelOption<T>
+/**
+ * Strictly descendant of a candidate class.
+ * 
+ * @author ejemba
+ */
+public class ClassDescendantOf implements Predicate<Class<?>>
 {
-    private String name;
-    private T value;
 
-    public KernelOption(String name) {
-        this.name = name;
+    private Class<?> ancestorType;
+
+    public ClassDescendantOf(Class<?> ancestorType)
+    {
+        this.ancestorType = ancestorType;
     }
-
-    public KernelOption(String name, T value) {
-        this.name = name;
-        this.value = value;
-    }
-
-    public KernelOption(KernelOption<T> option, T value) {
-        this.name = option.name;
-        this.value = value == null ? option.getValue() : value;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public T getValue() {
-        return value;
-    }
-
+    
     @Override
-    public String toString() {
-        return name + ": " + value;
+    public boolean test(Class<?> candidate)
+    {
+        return candidate != null &&  candidate != ancestorType && ancestorType.isAssignableFrom(candidate);
     }
+
 }