+++ tests/src/main/java/io/nuun/kernel/tests/internal/NuunITPlugin.java
@@ -19,16 +19,16 @@ package io.nuun.kernel.tests.internal;
 import io.nuun.kernel.api.plugin.InitState;
 import io.nuun.kernel.api.plugin.context.InitContext;
 import io.nuun.kernel.api.plugin.request.ClasspathScanRequest;
+import io.nuun.kernel.api.predicates.ClassAnnotatedWith;
 import io.nuun.kernel.core.AbstractPlugin;
 import io.nuun.kernel.tests.it.NuunITRunner;
 import io.nuun.kernel.tests.it.annotations.ITBind;
+import org.junit.runner.RunWith;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Map;
-
-import org.junit.runner.RunWith;
-import org.kametic.specifications.Specification;
+import java.util.function.Predicate;
 
 
 /**
@@ -41,7 +41,7 @@ public class NuunITPlugin extends AbstractPlugin
     private Collection<Class<?>> integrationTestsClass;
 
     @SuppressWarnings("unchecked")
-    private final Specification<Class<?>> iTSpecs = or(classAnnotatedWith(RunWith.class), classAnnotatedWith(ITBind.class));
+    private final Predicate<Class<?>> iTSpecs = new ClassAnnotatedWith(RunWith.class).or(new ClassAnnotatedWith(ITBind.class));
 
     @Override
     public String name() {
@@ -51,10 +51,10 @@ public class NuunITPlugin extends AbstractPlugin
     @Override
     @SuppressWarnings("rawtypes")
     public InitState init(InitContext initContext) {
-        Map<Specification, Collection<Class<?>>> scannedTypesBySpecification = initContext.scannedTypesBySpecification();
-        Collection<Class<?>> iTClassCandidates = scannedTypesBySpecification.get(iTSpecs);
+        Map<Predicate<Class<?>>, Collection<Class<?>>> scannedTypesByPredicate = initContext.scannedTypesByPredicate();
+        Collection<Class<?>> iTClassCandidates = scannedTypesByPredicate.get(iTSpecs);
         if (iTClassCandidates != null && !iTClassCandidates.isEmpty()) {
-            integrationTestsClass = new ArrayList<Class<?>>();
+            integrationTestsClass = new ArrayList<>();
             for (Class<?> itCandidate : iTClassCandidates) {
                 if (itCandidate.getAnnotation(RunWith.class) != null &&
 //                        itCandidate.getAnnotation(RunWith.class).value().equals(NuunITRunner.class)
@@ -69,7 +69,7 @@ public class NuunITPlugin extends AbstractPlugin
 
     @Override
     public Collection<ClasspathScanRequest> classpathScanRequests() {
-        return classpathScanRequestBuilder().specification(iTSpecs).build();
+        return classpathScanRequestBuilder().predicate(iTSpecs).build();
     }
 
     @Override